修改套餐信息

+++ src/main/java/com/reggie/controller/SetmealController.java
@@ -3,13 +3,12 @@ package com.reggie.controller;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.reggie.common.CustomException;
 import com.reggie.common.R;
 import com.reggie.dto.SetmealDto;
 import com.reggie.entity.Setmeal;
-import com.reggie.service.SetmealDishService;
 import com.reggie.service.SetmealService;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -19,11 +18,9 @@ import javax.annotation.Resource;
 @RequestMapping("/setmeal")
 public class SetmealController {
 
-    @Autowired
-    private SetmealDishService setmealDishService;
-
     @Resource
     private SetmealService setmealService;
+
 
     @GetMapping("/page")
     public R<Page<Setmeal>> page(int page, int pageSize, String name){
@@ -53,10 +50,33 @@ public class SetmealController {
         return R.success("状态更改成功");
     }
 
+    /***
+     * 回显数据
+     */
+    @GetMapping("/{id}")
+    public R<SetmealDto> getByid(@PathVariable long id){
+        log.info("id:{}",id);
+        SetmealDto byIdWithSetmeal = setmealService.getByIdWithSetmeal(id);
+        return R.success(byIdWithSetmeal);
+    }
+
+    /***
+     * 更新数据
+     */
+    @PutMapping
+    public R<String> update(@RequestBody SetmealDto setmealDto){
+        setmealService.updateSetmeal(setmealDto);
+        return R.success("更新数据成功");
+    }
     @DeleteMapping
     public R<String> delete(String[] ids){
         for (String id : ids) {
-            setmealService.removeById(id);
+            Setmeal setmeal = setmealService.getById(id);
+            if (setmeal.getStatus() != 0) {
+                setmealService.removeById(id);
+            }else {
+                throw new CustomException("状态不对");
+            }
         }
         return R.success("删除成功");
     }
